LayoutTests/http/tests/websocket/tests/hixie76/error-detect.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure WebSocket correctly fire error event for unknown frame type.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 1);

var errorCount = 0;

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hixie76/unknown-frame-type");
ws.onopen = function () {
    debug("WebSocket is open");
};
ws.onmessage = function (evt) {
    debug("received:" + evt.data);
};
ws.onerror = function () {
    errorCount += 1;
};
ws.onclose = function () {
    debug("WebSocket is closed");
    shouldBe("errorCount", "255");
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
